Package restructuring to enable deployment
