+++ LBpy/__main__.py
@@ -1,0 +1,5 @@
+# Executes the fluidLB solver as a script
+
+# if __name__ == '__main__':
+from .LBpy import main
+main()
