+++ run.py
@@ -1,0 +1,4 @@
+from LBpy.LBpy import main
+
+if __name__ == '__main__':
+    main()
